Append Presenter suffix to presenters
Much like with Repositories and Interfaces, it would be much easier to
navigate files in editors if it's clear whether you will be editing a
presenter or a model (ex: User model is User.php while its presenter is
also a User.php file) which can currently be ambiguous at times.

This commit also modifies the database models to reference the new
class names of their respective presenters.